RUBY-852 Refactor to ReadPreferrable module
