+++ lib/mongo/operation.rb
@@ -16,6 +16,7 @@ require 'mongo/operation/result'
 require 'mongo/operation/executable'
 require 'mongo/operation/specifiable'
 require 'mongo/operation/limited'
+require 'mongo/operation/read_preferrable'
 require 'mongo/operation/batchable'
 require 'mongo/operation/read'
 require 'mongo/operation/write'
