+++ spec/mongo/operation/write/bulk_update_spec.rb
@@ -520,10 +520,5 @@ describe Mongo::Operation::Write::BulkUpdate do
         expect(result.replies).to be(nil)
       end
     end
-
-    context 'when the server is a secondary' do
-
-      pending 'it raises an exception'
-    end
   end
 end
