+++ spec/mongo/operation/write/command/update_spec.rb
@@ -82,13 +82,6 @@ describe Mongo::Operation::Write::Command::Update do
 
     context 'server' do
 
-      context 'when the type is secondary' do
-
-        it 'throws an error' do
-          expect{ op.execute(secondary_context) }.to raise_exception
-        end
-      end
-
       context 'message' do
         let(:expected_selector) do
           { :updates       => updates,
