+++ spec/mongo/operation/write/delete_spec.rb
@@ -174,10 +174,5 @@ describe Mongo::Operation::Write::Delete do
         end
       end
     end
-
-    context 'when the server is a secondary' do
-
-      pending 'it raises an exception'
-    end
   end
 end
