+++ spec/mongo/operation/write/insert_spec.rb
@@ -227,10 +227,5 @@ describe Mongo::Operation::Write::Insert do
         end
       end
     end
-
-    context 'when the server is a secondary' do
-
-      pending 'it raises an exception'
-    end
   end
 end
