+++ spec/mongo/operation/write/update_spec.rb
@@ -173,10 +173,5 @@ describe Mongo::Operation::Write::Update do
         end
       end
     end
-
-    context 'when the server is a secondary' do
-
-      pending 'it raises an exception'
-    end
   end
 end
